@@ -1,11 +1,19 @@
-const sveltePreprocess = require("svelte-preprocess");
-module.exports = {
+import adapter from '@sveltejs/adapter-auto';
+import preprocess from 'svelte-preprocess';
+
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	// Consult https://github.com/sveltejs/svelte-preprocess
+	// for more information about preprocessors
 	preprocess: [
-		sveltePreprocess({
-			defaults: {
-				style: "postcss",
-			},
+		preprocess({
 			postcss: true
-		}),
+		})
 	],
+
+	kit: {
+		adapter: adapter()
+	}
 };
+
+export default config;
