@@ -1,16 +1,13 @@
 const tailwindcss = require('tailwindcss');
-const autoprefixer = require("autoprefixer");
-const cssnano = require("cssnano");
+const autoprefixer = require('autoprefixer');
 
-const mode = process.env.NODE_ENV;
-const dev = mode === "development";
-
-module.exports = {
+const config = {
 	plugins: [
-		tailwindcss,
-		autoprefixer,		
-		!dev && cssnano({
-			preset: "default",
-		}),
-	],
+		//Some plugins, like tailwindcss/nesting, need to run before Tailwind,
+		tailwindcss(),
+		//But others, like autoprefixer, need to run after,
+		autoprefixer
+	]
 };
+
+module.exports = config;
